[FIX][MTD VAT 8] Fixing an error during some reconcilation
To reconcile we need to create a balanced HMRC VAT Submission Journal
Entry. If the difference for the liability account_move_line was
negative we were getting the folowing error "Wrong credit or debit value in accounting entry !"
This was because when cretaing the move line the value contained the -
sign which needed to be removed. Also fixed a typo in comment