[IMP][MTD VAT V8] Creating new compute field.
Adding a new field to get the tax-amount for manually created Journal entries with Tax codes. As these Journal entries are not of the Purchase or sales type, the tax_amount does not get calculated. We will be using the new field to ge total value during the calculation of VAT. MTD specific field will be invisible for user.