[O1851] Return promise from outer 'then'
I got a sudden thought over the weekend that just returning the
promise from the inner 'then' might not be enough and that it was working
by accident with potential for race conditions, so I've had to produce
this mess instead.

I'm not sure it's strictly necessary to do this but I don't see it doing
any harm and it might just prevent a hard-to-find bug later on.
It stops the heebie-jeebies I had anyway.