[O1851] Remove nasty as_string hack.
It looks like it can cope with action.domain being set to a domain data
structure returned by Odoo rather than its string representation.

This works because we're not embedding client-side variables,
which is fine as we are resolving active_id explicitly in the JS just in time to call the
move_line_domain... method.  If we had wanted to keep any variables
client-side, then we would have had to keep it as a string, and when we
use it server-side we would also have to eval() with active_id set.

Thankfully we don't have to do such nastiness, because of the point at
which we're hooking in within the javascript.