[T15219] Fix expected singleton error caused by unstored search key
unique_number was computed but not stored, but later searched on
(I believe this is for reconcilitation of VAT, guessing from the
stacktrace I saw in the logs).

The reason it got to be a singleton error is because Odoo's ORM
ignores with a warning in the logs (rather than raising an exception)
when searching on fields that aren't searchable.  It would still have
been a crash but would have been a more immediately-useful one if it had
failed fast on that.