[O2010] Pass writes to 'reconcile' through to our backing field
I could have just replaced the field in the GUI, but
there may be places in other modules where the reconcile flag is set
via data files or via the ORM or via other forms, so this should allow
it to pass straight through.

It does mean there is an imbalance between write and read when the
context we use is set, but we're only going to be using that context in
a small corner to trigger it via code anyway.