[FIX][T15219] remving the search on api_id
This is so that we don't get into the 500 server error as When returned
to the server we do not have any means of knowing which api was used to
create the request. Originally the design was to allow one request tracker
per module but this can cause 500 server error as we can have two open request
trackers.