[REF] Make the target field names the keys, and boxes the values
It's customary in a lookup table such as this to put the name of the
field you're setting as the key, and the data driving what you're
setting it to as the value.

The reason this convention exists is because the fields we're setting
are guaranteed unique, but where the values come from not necessarily
so in the general case.  It's only be chance that we don't want to put
the same box figure in two different fields in our case, but if we had
it would have been impossible with the fields as values.